+++ src/main/java/com/hazelab/todo/controller/TaskController.java
@@ -1,34 +1,14 @@
 package com.hazelab.todo.controller;
 
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.servlet.ModelAndView;
 
-import com.hazelab.todo.entity.TaskEntity;
-import com.hazelab.todo.service.TaskService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
-
-@RestController
-@RequiredArgsConstructor
 public class TaskController {
-
-    private final TaskService taskService;
-
-    @GetMapping("/tasks")
-    public List<TaskEntity> findList(){
-        return taskService.findList();
-    }
-
-    @GetMapping(path="/tasks/{id}")
-    public Optional<TaskEntity> find(@PathVariable Integer id){
-        return taskService.find(id);
-    }
-
-    @PostMapping("/tasks/delete/{id}")
-    public ResponseEntity<String> delete(@PathVariable Integer id ){
-        taskService.delete(id);
-        return ResponseEntity.ok("success");
+    @GetMapping("task")
+    @ModelAttribute
+    public ModelAndView index(ModelAndView model){
+        model.addObject("/task");
+        return model;
     }
 }
