+++ src/main/resources/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <div class="content-top clearfix">
+        <h1 class="al-title">ご飯</h1>
+        <ul class="breadcrumb al-breadcrumb">
+            <li><a href="/task/">Home</a></li>
+            <li>ぽぽぽぽ</li>
+        </ul>
+    </div>
+</head>
+<body>
+
+<h1>ご飯だよ</h1>
+<h2>疲れた</h2>
+<h3>もうやめたい</h3>
+
+</body>
+</html>
